src/requirements.txt
azure-cosmos
azure-identity
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
flask[async]
python-dotenv
semantic_kernel